refactor: rename files following file naming convention (snake_case) and create a variable with supported langs instead of a getter function
