+++ internal/utils/find_similar_langs.go
@@ -4,7 +4,6 @@ import "strings"
 
 func FindSimilarLangs(language string) string {
 	var similarLangs []string
-	supportedLangs := supportedLangs()
 
 	for k := range supportedLangs {
 		for _, char := range language {
