+++ internal/utils/show_supported_langs.go
@@ -1,0 +1,15 @@
+package utils
+
+import "fmt"
+
+var supportedLangs = map[string]bool{
+	"go":   true,
+	"rust": true,
+}
+
+func ShowSupportedLangs() {
+	fmt.Println("supported languages:")
+	for k := range supportedLangs {
+		fmt.Println("  " + k)
+	}
+}
