--- internal/utils/supportedLangs.go
@@ -1,12 +1,0 @@
-package utils
-
-type supportedLanguages map[string]bool
-
-func supportedLangs() supportedLanguages {
-	languagesSuported := map[string]bool{
-		"go":   true,
-		"rust": true,
-	}
-
-	return languagesSuported
-}
