+++ internal/utils/verify_lang.go
@@ -6,9 +6,7 @@ import (
 )
 
 func VerifyIfLangIsSupported(language string) {
-	langs := supportedLangs()
-
-	_, ok := langs[language]
+	_, ok := supportedLangs[language]
 
 	if !ok {
 		fmt.Printf("The language '%s' is not supported\n", language)
